Update central-sync-release.yml
Updating back to bigger private runner

Signed-off-by: Chris Bono <chris.bono@broadcom.com>
Signed-off-by: Arno Geismar <arno.geismar@acerta.be>

Added pulsar binder support. In our organisation we only have a pulsar broker and want to avoid using a kafka protocol handler

Signed-off-by: Arno Geismar <arno.geismar@acerta.be>

Added pulsar binder support. In our organisation we only have a pulsar broker and we want to avoid using a kafka protocol handler to talk to pulsar

Signed-off-by: Arno Geismar <arno.geismar@acerta.be>

Added pulsar binder support. In our organisation we only have a pulsar broker and we wan't to avoid using a kafka protocol handler to talk to pulsar